fix: use string for addresses in queries
